Added `novel` library
