+++ package.json
@@ -11,20 +11,6 @@
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^4.16.2",
-    "@tiptap/core": "^2.0.3",
-    "@tiptap/extension-color": "^2.0.3",
-    "@tiptap/extension-horizontal-rule": "^2.0.3",
-    "@tiptap/extension-image": "^2.0.3",
-    "@tiptap/extension-link": "^2.0.3",
-    "@tiptap/extension-placeholder": "^2.0.3",
-    "@tiptap/extension-task-item": "^2.0.3",
-    "@tiptap/extension-task-list": "^2.0.3",
-    "@tiptap/extension-text-style": "^2.0.3",
-    "@tiptap/extension-underline": "^2.0.3",
-    "@tiptap/pm": "^2.0.3",
-    "@tiptap/react": "^2.0.3",
-    "@tiptap/starter-kit": "^2.0.3",
-    "@tiptap/suggestion": "^2.0.3",
     "@tremor/react": "^3.4.1",
     "@upstash/ratelimit": "^0.4.3",
     "@vercel/analytics": "^1.0.1",
@@ -40,9 +26,10 @@
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.244.0",
     "nanoid": "^4.0.2",
-    "next": "canary",
+    "next": "13.4.20-canary.15",
     "next-auth": "4.22.1",
     "next-mdx-remote": "^4.4.1",
+    "novel": "^0.1.16",
     "openai-edge": "^1.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -53,8 +40,6 @@
     "sonner": "^0.4.0",
     "swr": "^2.2.0",
     "tailwind-merge": "^1.13.2",
-    "tippy.js": "^6.3.7",
-    "tiptap-markdown": "^0.8.1",
     "unist-util-visit": "^4.1.2",
     "use-debounce": "^9.0.4"
   },
@@ -72,7 +57,7 @@
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.3.0",
     "prisma": "^4.16.2",
-    "tailwindcss": "^3.3.2",
+    "tailwindcss": "^3.3.3",
     "tailwindcss-animate": "^1.0.6",
     "typescript": "^5.1.6"
   }
